.vim: js: iabbrev rewrite
At work we require a space before the opening paren of an anonymous JS
function. In my own code I don’t. In wp-calypso there are spaces on the
inside of parens, similar to jQuery.

Because of this I had to have a per-project javascript.vim which
re-defined of all the iabbrevs that contained anonymous functions, like
fn, fnus and te(de|be|af|it). Not good.

This new code allows having the function literal defined in a separate
“private” iabbrev -- _fn -- and then reused everywhere. So, now I will
have to only redefine _fn in my per-project javascript.vim.