Give up on msmtp
For some reason it worked perfectly when I interactively issued the mail
command from shell, but did NOT from crontab, which actually was the
main use-case for setting up email in the first place.

I have figured out how to setup Postfix to use my Gmail account, so now
it works system-wide.